Consolidated configuration in one file
