+++ app.js
@@ -2,11 +2,10 @@ var mongoose = require('mongoose');
 var express = require('express');
 var bodyParser = require('body-parser');
 var app = express();
-var db = require('./backend/config/db');
-var server = require('./backend/config/server');
+var config = require('./backend/config/config');
 
 //Database
-mongoose.connect(db.url);
+mongoose.connect(config.db.url);
 
 //Middleware
 app.use(express.static(__dirname + '/public'));
@@ -16,6 +15,6 @@ app.use(bodyParser.urlencoded({extended:true}));
 //Routing
 require('./backend/routes/routes')(app);
 
-app.listen(server.port, function() {
-  console.log('Server listening on port ' + server.port);
+app.listen(config.server.port, function() {
+  console.log('Server listening on port ' + config.server.port);
 });
