--- backend/config/db.js
@@ -1,13 +1,0 @@
-var db = {};
-//TODO Common configuration
-switch(process.env.NODE_ENV) {
-  case 'dev':
-    db = {url: 'mongodb://localhost/dev'};
-    break;
-  case 'prod':
-    //TODO Add security
-    db = {url: 'mongodb://localhost/lanmomo'};
-    break;
-}
-
-module.exports = db;
