--- backend/config/mail-options.js
@@ -1,4 +1,0 @@
-module.exports = {
-  from: 'Lan Montmorency',
-  subject: 'Vérification de courriel'
-};
