--- backend/config/server.js
@@ -1,11 +1,0 @@
-var server = {};
-switch(process.env.NODE_ENV) {
-  case 'dev':
-    server = {port: 3000};
-    break;
-  case 'prod':
-    server = {port: 80};
-    break;
-}
-
-module.exports = server;
